refactor(tooling): centralize artifact generator contracts

Copilot-Session: 642dd193-e81f-4e2a-a976-e20c93945efd
